refactor: migrations conditions

+++ alexa-rpg-backoffice/src/infra/db/migrations/1734211141521-add-tags-to-segment.ts
@@ -3,19 +3,26 @@ import { MigrationInterface, QueryRunner, TableColumn } from 'typeorm';
 
 export class AddTagsToSegment1734211141521 implements MigrationInterface {
   public async up(queryRunner: QueryRunner): Promise<void> {
-    await queryRunner.addColumn(
-      'segment',
-      new TableColumn({
-        name: 'tags',
-        type: 'varchar',
-        isNullable: false,
-        isArray: true,
-        default: 'ARRAY[]::varchar[]',
-      }),
-    );
+    const exists = await queryRunner.hasColumn('segment', 'tags');
+
+    if (!exists) {
+      await queryRunner.addColumn(
+        'segment',
+        new TableColumn({
+          name: 'tags',
+          type: 'varchar',
+          isNullable: false,
+          isArray: true,
+          default: 'ARRAY[]::varchar[]',
+        }),
+      );
+    }
   }
 
   public async down(queryRunner: QueryRunner): Promise<void> {
-    await queryRunner.dropColumn('segment', 'tags');
+    const exists = await queryRunner.hasColumn('segment', 'tags');
+    if (exists) {
+      await queryRunner.dropColumn('segment', 'tags');
+    }
   }
 }
