+++ alexa-rpg-backoffice/src/infra/db/migrations/1734211947077-add-tags-to-action.ts
@@ -2,19 +2,27 @@ import { MigrationInterface, QueryRunner, TableColumn } from 'typeorm';
 
 export class AddTagsToAction1734211947077 implements MigrationInterface {
   public async up(queryRunner: QueryRunner): Promise<void> {
-    await queryRunner.addColumn(
-      'action',
-      new TableColumn({
-        name: 'tags',
-        type: 'varchar',
-        isNullable: false,
-        isArray: true,
-        default: 'ARRAY[]::varchar[]',
-      }),
-    );
+    const exists = await queryRunner.hasColumn('action', 'tags');
+
+    if (!exists) {
+      await queryRunner.addColumn(
+        'action',
+        new TableColumn({
+          name: 'tags',
+          type: 'varchar',
+          isNullable: false,
+          isArray: true,
+          default: 'ARRAY[]::varchar[]',
+        }),
+      );
+    }
   }
 
   public async down(queryRunner: QueryRunner): Promise<void> {
-    await queryRunner.dropColumn('action', 'tags');
+    const exists = await queryRunner.hasColumn('action', 'tags');
+
+    if (exists) {
+      await queryRunner.dropColumn('action', 'tags');
+    }
   }
 }
